requirements.txt
streamlit
PyMuPDF
sentence-transformers
faiss-cpu
python-dotenv
